api/app/comment/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import Comment
from .serializers import CommentSerializer

class CommentViewSet(viewsets.ModelViewSet):
  queryset = Comment.objects.all()
  serializer_class = CommentSerializer

  def perform_create(self, serializer):
    serializer.save(owner_id=self.request.user)